backend/requirements.txt
Django==4.1.7
djoser==2.1.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
drf-extra-fields==3.4.1
Pillow==9.4.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
pdfkit==1.0.0
